Update linkedlist_project
This is a database is for Hot Water which is a small start-up company that sells spas. Hot Water does not carry any stock. A few spas are set up in a simple warehouse so customers can see some of the models available, but any products sold must be ordered at the time of the sale.